Finished Live project
This project is an android application developed for Suvishesha Bhavan, Cate- chetical Centre under the Diocese of ********. This project aims to create an application to work in tandem with the existing website of the said client. There will be one main admin app under the Diocese of ******* who can add promoter, open registration portals for exams and kalolsavam events. And second app ,each promoter(sunday school centres)  can register students on portals. The application also aims to provide timely notiﬁcations for exams and seminars to be held and that of results published.
The students can use to see the results of exams and kalolsavam. The audit process for the sunday schools can be done through this app,currently it is done through paperwork. It consists of two app one for admin and another for sunday school centres and also students.